fix(front): fix unwanted theme change on opening filter settings modal
remove debug border from filters modal

move theme detector from component to store
