+++ web/src/ui/components/FilterModal/FilterModal.scss
@@ -1,5 +1,4 @@
 .FilterModal {
-  border: 1px solid red;
   .modal.modal-blur.fade.show {
     padding-right: 1rem;
     display: block;
