+++ web/src/ui/components/ThemeToggler.js
@@ -1,22 +1,8 @@
-import React, {useContext, useEffect} from 'react';
+import React, {useContext} from 'react';
 import {ThemeContext} from '../../store/ThemeStore';
-
-const detectBrowserTheme = () => {
-  const supportsPreference =
-    window.matchMedia('(prefers-color-scheme)').media !== 'not all';
-  const isLight =
-    supportsPreference &&
-    window.matchMedia('(prefers-color-scheme: light)').matches;
-  return {isLight};
-};
 
 const ThemeToggler = () => {
   const {theme, changeTheme} = useContext(ThemeContext);
-
-  useEffect(() => {
-    const {isLight} = detectBrowserTheme();
-    if (isLight) changeTheme('light');
-  }, []);
 
   return (
     <div
